requirements.txt
Flask
Flask-SQLAlchemy
Flask-Bcrypt
PyMySQL
PyJWT
python-dotenv
Pillow
matplotlib